+++ src/main/java/de/alpharogroup/db/entity/delete/Deletion.java
@@ -75,6 +75,12 @@ public class Deletion<PK extends Serializable, T, U> extends BaseEntity<PK>
 		IdentifiableDeletable<PK, T, U>
 {
 
+	/** The Constant for the column name 'deleted'. */
+	public static final String COLUMN_NAME_DELETED = "deleted";
+
+	/** The Constant for the column name 'deletedBy'. */
+	public static final String COLUMN_NAME_DELETED_BY = "deletedBy";
+
 	/** The Constant serialVersionUID. */
 	private static final long serialVersionUID = 1L;