+++ src/main/java/de/alpharogroup/db/entity/name/NameEntity.java
@@ -51,6 +51,9 @@ import lombok.experimental.FieldDefaults;
 public abstract class NameEntity<PK extends Serializable> extends BasicNameEntity<PK>
 {
 
+	/** The Constant for the column name 'name'. */
+	public static final String COLUMN_NAME_NAME = "name";
+
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;