+++ src/main/java/de/alpharogroup/db/entity/text/TextEntity.java
@@ -54,6 +54,9 @@ public abstract class TextEntity<PK extends Serializable> extends BaseEntity<PK>
 		IdentifiableTextable<PK>
 {
 
+	/** The Constant for the column name 'text'. */
+	public static final String COLUMN_NAME_TEXT = "text";
+
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;