+++ src/main/java/de/alpharogroup/db/entity/version/VersionableBaseEntity.java
@@ -40,7 +40,10 @@ import lombok.Setter;
  *
  * @param <PK>
  *            the generic type of the id
+ * @deprecated use instead the new renamed class <code>VersionableEntity</code><br>
+ *             Note: will be removed in the next minor release
  */
+@Deprecated
 @MappedSuperclass
 @Getter
 @Setter